+++ config/postcss.config.js
@@ -3,9 +3,9 @@ const rootPath = path.join(__dirname, "..");
 
 module.exports = {
   plugins: [
+    require("postcss-preset-env"),
     require("postcss-import"),
     require("postcss-mixins"),
-    require("postcss-preset-env"),
     require("cssnano"),
     require("postcss-url")({
       url: "copy",